classifier-service/requirements.txt
fastapi
uvicorn[standard]
pika==1.3.2
psycopg2-binary
prometheus_client

